Make unit controls accessible with a keyboard
Before this commit input radio buttons where hidden from the render
tree and the only thing the user could access to control the state
of the form was labels, that's not a good UX for people who navigate
the website with a keyboard. That's why I brought the radio buttons
back to the render tree but made them "invisible".