Remove the default value from TemperatureToggle
One of the expectations about TemperatureToggle component and one of its
test cases were invalid. The 'selectedUnit' prop is required
on TemperatureToggle controlled component in order to make it work
and to keep React "the single source of truth". It shouldn't have
any default values, neither should it be optional, the application
is in full control of what field is checked.