Add status code check of fetched data
Application didnt have any sanity checks preventing it from crashing
if something goes wrong. For this reason I added an if statement
to check whether the data has arrived with the 200 status code.
I'm not using a strict equality operator because the API I'm using
returns a 'cod' property as a string or as a number depending on the
API endpoint.