Remove fetchAutocomplete errors
A simple fix to two errors that were occuring in the console.
Firstly, before this commit queryFn didn't return any value.
According to the definition of the QueryFunction type it should
return either the Promise or the generic. For that reason I used
the fetch call that was already there as the returned Promise.

Second problem was one of the sanity checks assumes that statusCode
property is always inside of the response with autocompletion data
and of course it's not the case. The statusCode property appears
only in error responses, hence I added a simple check.